requirements.txt
PyYAML==6.0
ShopifyAPI==10.0.0